feat: implement lazy loading for components and add new charts for dashboard analytics
