+++ src/redux/creditCardsSlice.ts
@@ -1,23 +1,10 @@
 import { createSlice } from "@reduxjs/toolkit";
 import type { PayloadAction } from "@reduxjs/toolkit";
 import { fetchCreditCardData } from "./operations";
-
-export interface CreditCard {
-  id: number;
-  userId: number;
-  cardNumber: string;
-  cardType: string;
-  cardHolder: string;
-  validThru: string;
-  balance: number;
-  theme: string;
-}
-
-export interface CreditCardsState {
-  cards: CreditCard[];
-  loading: boolean;
-  error: string | null;
-}
+import type {
+  CreditCard,
+  CreditCardsState,
+} from "../types/creditCardsSliceTypes";
 
 const initialState: CreditCardsState = {
   cards: [],
