+++ src/styles/Security.module.css
@@ -40,7 +40,7 @@
 .pass__title {
   color: rgb(51, 59, 105);
   font-size: 17px;
-  font-weight: 500;
+  font-weight: 600;
   line-height: 21px;
   letter-spacing: 0px;
   text-align: left;
