+++ src/types/transactionsSliceTypes.ts
@@ -1,0 +1,16 @@
+export interface Transaction {
+  id: number;
+  userId: number;
+  description: string;
+  transactionDate: string;
+  amount: number;
+  type: string;
+  status: string;
+  icon: string;
+}
+
+export interface TransactionsState {
+  transactions: Transaction[];
+  loading: boolean;
+  error: string | null;
+}
